requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
attrs==25.3.0
Authlib==1.5.2
av==14.4.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==44.0.3
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
decord==0.6.0
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==8.0.2
et_xmlfile==2.0.0
fastapi==0.115.12
faster-whisper==1.1.1
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gradio==5.29.1
gradio_client==1.10.1
greenlet==3.2.2
groovy==0.1.2
h11==0.16.0
helium==5.1.1
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
litellm==1.70.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
openai==1.75.0
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
primp==0.15.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pybaseball==2.2.7
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==3.2.3
PySocks==1.7.1
python-chess==1.999
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.0
ruff==0.11.10
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.3
selenium==4.32.0
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smolagents==1.16.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.41
stable-ts==2.19.0
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
Transcriber==0.1.19
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.0
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
Wikipedia-API==0.8.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
yt-dlp==2025.4.30
zipp==3.21.0
zstandard==0.23.0
